Rename test to support travis
The test has been renamed to support the travis test suite (one of the
builds is failing). The test code should really be added to the existing
test_astro_session.py test, but it's not clear yet what the best approach
is here (since there is the replication of the tests from the
sherpa.ui.utils.Session object).