Separate out test for contents of list_samplers
The contents of the list_samplers return value is now separated out into
a separate test. As this routine is also available from the ui layer new
tests are added for it (essentially copies of the same test).

I have created new test files for the "ui" versions of the tests: these
may be better in the session tests (but see discussion of session tests
in #396) or the exsiting "unit" test for the ui code (e.g.
test_astro_ui_utils_unit.py)