Always show warnings. Drop --show-warnings.
I have found that I always want to show warnings. The display has been
changed to only show "new" warnings, since I have seen times when the
same warning has been reported many times, which is not so useful. This
code could report the number of times a warning is repeated, but I
don't htink it is worth it at this time; the total number of warnings
and a counter are displayed so you can see that there are repeated
warnings, if necessary.

Another reason for removing --show-warnings is that I have recently
been unable to get it to work, and have seen failures on Travis due
to it (only when a warning had been created; it wasn't breaking
successful tests).