Normalize the docstrings for deprecated/aliased parameters
There were two sections in each docstring where the parameter aliases were
mentioned: a ..note section (styled on the AstroPy guidelines) in the
summary text, and a Notes section before the References section. This
commit removes the Notes section, retaining only the AstroPy-suggested
..note section. The text for this section was also re-worded (although
it still is not ideal). The layout of the ..note section does not quite
match the AstroPy guideline - I have added a blank linke between the note
and the contents, and have aligned the text to the note rather than the
title of the note - since these produce more readable text and better
Sphinx output (when not using the AstroPy styling).

An alternative would have been to use the Sphinx ..deprecated directive,
but I decided to stick with the AstroPy-suggested format for now.

I also fixed up several issues (e.g. models which claimed to have a parameter
alias but did not, a failure to use the new parameter name in the text,
using 4.9.0 rather than 4.10.0 as the version). I ran a check to validate
that only models with a parameter alias now mention these aliases.