Add XSPEC 12.9.1 model slimbh
This is the first time we interface to an XSPEC model labelled as "c_"
(i.e. a lower-case C). It appears to work. Do we need to look at the
interface to "C_" models and see if any tweaks could be made, given
that we appear to be handling two different APIs with the same template.

The leading "c_" is dropped to get the linking to work. This means that
it is harder for a user to distinguish, when inspecting the XSPEC class
contents, between single-precision Fortran style routines and this one.
Is this something that we want to look into in the future?