Workaround XSPEC ismabs parameter name clashes
Use the full names for Si and S parameters - that is SiliconI,
SiliconII, SiliconIII, SulfurI, SulfurII, and SulfurIII - rather
than SiI, SiII, SiIII, SI, SII, SIII which conflict.

A test has been added to check these, to act as documentation and a
check for if we ever move to an auto-generated XSPEC module.

Perhaps only one of these should be named in full, with the other
one left as is?