Support new function calls for models from XSPEC 12.9.1
This commit changes the compiled code so that it will use the
correct version of the models when built against XSPEC 12.9.0 or
12.9.1.

However, the Python models will only work with 12.9.1, since I
have not worked out how to make the function names depend on the
XSPEC model library. It is probably easy to do but I haven't had
time to look at it.