Work around failure cases when backscal is an array
For grating PHA data sets the background scale value can vary with bin,
and so is returned as an array. This causes the creation of the
BackgroundSumModel instance - which is created whenever a model instance
is being created by sherpa.astro.ui functions like get_model and show_all -
to fail with a TypeError, as '%g' % backscale errors out.

This change traps this error and replaces the actual scaling value by
a token ("bkgscale"), which lets the process proceed. It is unclear
however what the semantics of the string representation of a model is:
is it purely for display purposes, in which this replacement is
okay, or is it meant to be actionable/machine readable (e.g. when
saved via the ui.save_all function)?