Change from error to warning for OGIP violations
Given that users can not easily change a response file, and previous
versions of Sherpa would allow the responses to be used, this commit
changes some of the errors recently introduced (in PR #383) into
warnings. The errors for the first bin edge being <= 0 are still
left in because users of the sherpa.astro.ui module will find that
these files are auto-corrected for them (by PR #383).