Support building Fortran extensions with conda gfortran
It appears that the LDFLAGS set up by the conda environment when it
installs gfortran_linux-64 breaks the Fortran build because NumPy's
distutils code takes the LDFLAGS setting wholesale, rather than
appending them to its "current" settings. See
https://github.com/numpy/numpy/issues/1171

The LDFLAGS setting could be removed before calling the setup code,
but it may be useful elsewhere in the code, so hardcode the addition
of "-shared" to the LDFLAGS environment variable, but only if set to
a non-empty value. This feels hacky; let's see how it works out.

The suggestion is based on feedback from Thomas Holder to a question
posed to the anaconda list - see
https://groups.google.com/a/continuum.io/forum/#!topic/anaconda/Xw57CjIcBIU
but all mistakes and confusion should be directed at the author of this
commit.