Initial support for XSPEC 12.10.0 (fix #436)
This adds the following XSPEC models (all additive) when built against
XSPEC 12.10.0:

    XSbrnei, XSbvrnei, XSbvvrnei, XSgrbcomp, XSjet, XSssa, XSzcutoffpl

The following models now use C/C++ rather than the FORTRAN interface,
so may result in numerical differences (when using XSPEC 12.10.0 or
later):

  XSc6mekl, XSc6pmekl, XSc6pvmkl, XSc6vmekl, XSnsmax, XSnsmaxg, XSnsx,
  XSacisabs, XSswind1, XSzxipcf

The low-level interface to the gsmooth and lsmooth convolution models
has been updated to use the updated function names for XSPEC 12.10.0
or later.

The documentation has been updated to note explicit support for XSPEC
12.10.0, and the old advice of how to build against the CIAO build of
XSPEC has been removed as it was not founf to be that useful.

The tests have been updated to match the 12.10.0 behavior; this is
primarily increasing tolerances for the models which have changed from
FORTRAN to C/C++ interfaces. The tolerances were adjusted until they
passed on a Linux machine; they were not checked to make sure they
still made sense.

This has only been tested on Linux and it would not be surprising if the
build fails on OS-X, since the steps needed to call the xs_getVersion
function from the XSPEC library are much-more involved.