Adjust XSPEC parameter names
XSPEC 12.10.0 was released with two parameter names that can not be
mapped directly to Python: log(A) for grbcomp and log(mdot) for jet.
The XSPEC team plan to rename them to log_A and log_mdot (as done
with other parameters), so switch to this naming scheme.

The xilconv and rfxconv models contain log(xi) parameter names, but
we currently do not provide a Python class for these models, so are
not an issue for this update.