Improve support for builds without a region library
The Region and region_mask symbols are only exported from sherpa.astro.utils
when the sherpa.astro.utils._region module is built, but this case was not
being considered by sherpa.astro.data.

Since the symbols could be used when un-pickling a state file, create
a warning about missing support rather than raising an ImportErr (when
it makes sense).

There are no tests for this, and it's not clear if we can still build
Sherpa without a _region module, so an alternative would be to just
make it a required part of the Sherpa installation and remove this
conditional code.