Revert tolerance change for a test
It looks like the reason for the tolerance change may not be XSPEC but
some other change (as this test fails on my machine wuing XSPEC 12.9.1p).
For now revert the change (to check it is un-needed for Travis).

I have done a limited check to see if the difference is due to NumPy
(e.v. 1.14 vs 1.12) but it does not seem to me. It could be a compiler
difference, or some other environmental change.