Add Python docstrings for sherpa.utils._psf routines
The docstrings are added to sherpa/utils so that they can be picked
up by Sphinx when _psf has not been built, to make them easier to
edit, and improve the user experience by having accurate signatures.

This is a first-go around on the documentation as it is not clear I
have understood everything about these routines.