Add a Sphinx build
The aim is to support builing the documentation without having to
build Sherpa, as that appears to be the easiest way to get the
documentation built on Read The Docs. The documentation build is
Python 3.5 (or later) only: it is unclear whether this is due to
the complicated import structure in Sherpa or changes to how the
mock object works in Python 2.7 compared to 3.5.

The auto-summary templates are taken from the AstroPy helpers project.

Note that the documentation for the optimisation section is a rework
of the material from the CIAO 3.4 version of the Sherpa manual, in
particular the memo from Mark Birkinshaw, which can be seen at
http://asc.harvard.edu/sherpa3.4/documents/manuals/html/refmethods.html

Unfortunately the optimisers have changed since this memo:
  a) there's a lot fewer
  b) the scatter-shot versions have been completely rewritten
so it's not clear if the advice section is still sensible.

When building with Sphinx, there are a large number of warnings like
the following:

    /home/djburke/sherpa/sherpa-add-sphinx-docs/sherpa/ui/utils.py:docstring
    of sherpa.ui.utils.Session.confidence:163: WARNING: Duplicate explicit
    target name: "1".

Is this because the same docstring is being re-used for multiple functions
(due to the inclusion of *.ui and Session object), an issue with the
documentation format (I have looked but don't see it), some other issue,
or a problem with Sphinx? It does mean that we can not run a "clean" build
of the documentaition in Travis.

One choice made in this documentation is not to auto-generate the
output (since the aim is to be able to build the documentation without
having to build Sherpa). There are scripts provided to generate the
screen output (and PNG files), but this is not complete.

Update the fit/simple-fit thread documentation

This removes reference to a bug which has been fixed in Sherpa 4.10.0
(the need for hline in a resid plot), as well as fixing the documentation
to address some errors. Added scripts to replicate the documents.