Build: comment out the extra_link_args for passing through F77 link args
This comments out the recent change which was to support easy building
of Sherpa using the conda-supplied gfortran. This was added to help
out the Read The Docs build but that approach has since been abanded (i.e.
the need to build Sherpa on Read The Docs).

The change to pass through the F77 link arguments is controversial,
since @olaurino feels it is error prone. It is also not clear if it
leads to breaking the OS-X Travis CI build. This commit comments out
the application of the extra arguments to the code that builds the
extensions to see if this does cause the OS-X problem.