add 1d plotting and adjust unit tests (see below though)
Some unit tests are still failing. I am not sure whether the tests should pass
to begin with.

The tests compare the model evaluation directly, with different grids. Since
the grids have different binning, it does not surprise me that the results are
different, and the values seem to make sense to me. But since the original tests
by @DougBurke expected the values to be the same I am assuming I am missing something
and they really should.

One test is also failing for more subtle reasons that may lie in one of the
comments I found in the code, i.e.
                #       Note that test_regrid1d_int_flux
                #       appears to fail if the grid width used for modelfunc
                #       is larger than the output.