Tweak the eqwidth docstring
This removes an erroneous piece of text which messed up the docstring
parsing in the Sphinx documentation, adds new lines to break up the
sections, and rewrites the return section slightly (it's not obvious
from the astropy/NumPy docstring "rules" what the best way to document
routines that return a variable number of arguments).

Also took out un-needed \ lines and updated the copyright year.