Add a simple DataPlot test with PHA data
This tests out the handling of the error calculation for data
objects. At present the only test is that the plot can be
prepared for each statistic value. The primary motivation for
this test is because the error calculation is known to be
wrong for non-Chi2XXX statistic classes due to a recent
change from the Sphinx documentation. This commit adds a simple
test which can validate any change.