Switch pileup test from unittest to pytest style
There is no significant functional change here, except that

a) the covariance check is moved later, so we can see whether the
   best-fit parameter values have changed (given the test
   tolerance)

b) the logging has been left at tis default level, which provides
   more screen output on failure.

The old unittest version has been left in, but renamed to _test_pileup
so that it is ignored.

There may be some subtle differences due to how the Sherpa "state"
is saved by the test_threads class compared to the clean_astro_ui
fixture. The difference is in whether the ui._session dict is
saved/restored, or whether sherpa.astro.ui.clean() is called.