Docs: remove trailing .dirty from revision on read the docs
The Sherpa revision can look like 'Sherpa 4.10.1+72.g066205f.dirty',
and this just removes the ".dirty" suffix (which indicates that
there have been changes to the git repository since the last commit).
This is seen on the read-the-docs build (the build process must
change or add something causing the .dirty version to appear).

This change is only made when run on read-the-docs (well, when the
READTHEDOCS environment variable is set to "True") as it can be
useful for local builds to note that something has changed in
the repository (e.g. changes have not been committed).