Docs: Use sphinx_astropy.ext.edit_on_github
Adds "edit on github" links to class/method/functions to take users
to the docstring (on GitHub) if they want to make a simple edit.

This doesn't always work - for instance overloaded methods which
do not have a docstring are shown, rather than the class which
contains the docstring, and we manage to confuse things so it
doesn't know where to go when looking at sherpa[.astro].ui (you
need to look at the sherpa[.astro].ui.utils.Session object).