Add a test that the XSPEC models evaluate to a positive value
This is to catch the XSPEC 12.10.0 case where the APEC models use
ATOMDB version 3.0.7 by default, but the data files are actually
version 3.0.9 (and so the models return 0).

To support the XSgaussian model, the default grid now ends at 7 keV
not 5 keV (since the model has no signal at 5 keV with the default
parameter settings of 6.5 keV with sigma=0.1 keV.

Several models actually end-up evaluating to 0 when using the default
parameters, so they are skipped. They should be investigated, but
that is a different issue.