Safely support XSPEC table models in XSPEC 12.10.1
In XSPEC 12.10.0 we found that using XSPEC table models could crash
Python, since the routines we used from XSPEC (xsatbl and xsmtbl)
were using an output stream that was not set up by FNINIT. We have
been given a patch to XSPEC 12.10.0 for CIAO, but it does not
help external users.

XSPEC 12.10.1 has provided a new interface to table models (tabint)
which subsumes xsatbl, xsmtbl, and xsetbl [the latter is not
supported by Sherpa]. This commit switches over the handling of
XSPEC table models to this new interface.

Note that, as of XSPEC 12.10.1b, this new interface is only proved
for FORTRAN and C++. This commit uses the C++ interface, but it
is expected that a future XSPEC patch will provide a C interface
which we will then switch to (at this point we would then want to
be able to say "don't use these versions of XSPEC", but there is
no nice way for us to do this). However, that is for the future.