Change to XSPEC test
Ensure that XSkerrdisk model is not skipped from the "must produce
a bin with a value > 0" test. This is needed because the unittest
version of the "check results are sensible" test (assertFinite) does
not have a clear meaning for the model argument, and so has been sent
several types for this argument (e.g. model class but also model name).
Ideally the tests will be moved over to pytest style tests, where
assert_is_finite is sent the class and so can be more precise in the
check.