Consistent calculation of rstat and qval
The calculation of rstat and qval is now done in one place, as previously
the logic used by FitResults was slightly different to that used by
Fit.calc_stat_info. Both values are now:

 - both None, when the statistic does not support a "goodness of fit"
 - both NaN, when a value is invalid (statistic < 0 or dof <= 0)
 - both numbers

Prior to this change, FitResults would return None for qval if rstat
was NaN, and calc_stat_info would return qval=1 for the same situation.