Simplify the README
As we now have a Read The Docs instance with the build instructions,
streamline the README so that it gives a minimal amount of advice and
then says "Look over here for more details".

I have been rather cavalier in deleting information from the README,
in that I have not carefully checked what information is not included
in the Sphinx documentation.