Add documentation on how the XSPEC library is initialized
This was added in order to provide some information on the change in
behavior introduced in XSPEC 12.10.1 - namely that the "models library"
interface that Sherpa uses now picks up settings from the user's
XSPEC configuration file. As the default for the cross-section table
has changed from bcmc to angr in this file, this means that users
are going to see differences from fit results when they switch to 4.11.0
with XSPEC 12.10.1 (for models that use the cross-section tables, which
such as some of the absorption models).