use DataIMG as the PSF kernel
Data2D does not have WCS information, so I am assuming that
it is ok to consider the pixel size to be the same. This should
be the equivalent of a regula application of kernel in image
manipulation. When working with photographs there is no such
concept as a "pixel size" and the kernle is just applied
pixel-by-pixel.

So I changed the test to use DataIMG instead of Data2D.

The tests are failing for the same reason. However, now
there is a warning issued because the PSF we created
does not have any pixel size information.