Initial test of plot warnings
Add several tests to check that we are creating the "The displayed
errorbars have been supplied with the data or calculated using
chi2xspecvar; the errors are not used in fits with ..." warning messages
when creating plots.

These tests do not require a plotting backend. However, given that the
behavior is "stateful", in that the tests could fail if some other
part of the system has changed the plot preferences, then there are
additional checks (beyond those needed to perform the checks) that
indicate whether the system state has been changed.

The caplog fixture, introduced in pytest 3.3, is used to capture the
logging output (which is how the warnings are displayed to the user;
they do not use the warnings module). There are other tests in the
system that change the system logging level, so these tests use the
caplog context mamanger to ensure the tests are run with the standard
logging level.