Internal debugging change only
I can not get the test to fail locally, but it reliably does on Travis.
Looking at where the error occurs, it looks like the check in
JointPlot.plottop, to see if the input is an instance of the fit plot,
isn't being triggered. Can we tell why from the test itself?