Fix plot-test-warning failures
This required changing an isinstance check in the plottop method of
JointPlot. It is not at all clear why the existing test was failing,
but running some previous test causes this test to fail, so I have
switched to a more-Pythonic approach (as the old code was labelled
FIXME), of checking for attributes rather than requiring a particular
class. The current solution is not much better, but it would likely
require a significant review to see how to re-engineer the plot
structures (such as FitPlot) to improve the situation.

There is at least one other area where an isinstance call is made
to check whether a plot is a "FitPlot", in the ui code, but this is
not changed here.