Minor code clean up
Manually spacing, but adds explict catches for two try/catch blocks,
and notes issues with one of these (potential loss of information in
the case of an unexpected failure, as well as confusion of < or <= 0),
and a note that this changes the global NumPy error-reporting state.

There should be no functional changes (i.e. it is expected that the
code has not changed behavior).