Fix #621 for data-only plots
This stops data plots from displaying a warning if the y error bars are
not to be displayed.

It is not clear from the existing code what the requirements on the
yerr attribute are: is it always expected to be set even when not
used, or is it okay for it to be null/empty array/whatever the to_plot
method on the data object returns?

Also note that there are two other plot types that can generate this
warning, namely ResidPlot and RatioPlot, but these are not changed
by this commit (as there's no tests). The code for these plots is
subtly different to the DataPlot (in particular the extra check for
a failure when calculating the errors, adding in a warning message
to the user, is not present in these but is in DataPlot), which
suggests a review of the behavior here is needed.