scale data before rebinning + other minor changes
more counts where introduces by evaluating the model on a bigger grid
without scaling the results to account for that. In principle we would
be better off integrating the models, but that is not possible because
some of the models are not available as integrated functions.

I also updated PSFSpace2D to produce a better spacing between the pixels
as well as decoupling the x vs y pixel size.

Finally, when the PSF has a worse resolution than the data, Sherpa
errors out.

Tests were updated to pass now.