datastack:show_stack now looks for MJD-OBS then MJD_OBS
The show_stack routine is purely informational, in that it
just displays information to the user about the loaded data sets.
It used the MJD_OBS keyword in the file to indicate the
observation time, as an optional value.

This has been updated to look for MJD-OBS and then, if that doesn't
exist, use MJD_OBS.