Remove python 2.7 support from the build support
The setup.py script now fails if run with a python older than 3.5.

The package metadata has been updated to reflect that Sherpa is now
3.5 or later (with python_requires set to not include Python 4 or
later), so that pip installations will only work with a supported
Python version.

It is not clear from
https://packaging.python.org/guides/dropping-older-python-versions/#dropping-a-python-release
whether we really should have included python_requires in the 4.11.1
release, but I think we are okay as is.

The Travis tests should fail because I have not removed the
python 2.7 build.