matplotlib: explicitly set zorder based on plot structure
Attempt to set points and error bars with a sensible zorder so that
subsequent plots (overlaid on this one) get displayed sensibly. This
needs a significant amount of usability testing since it changes how
the plots behave (in particular how the points and error bars are
displayed with respect to each other).

There is also a deprecation warning displayed if you use matplotlib 3.1 or
later, which I have added a comment about but not looked at.