XSPEC: build against 12.10.1m by removing support for models
Remove support for XSrelline, XSrelline_lp, and XSrelline_lp_ext models.
These were added in 12.10.1 but removed in 12.10.1m. It does not seem worth
supporting just those versions where they were available.

Remove support for low-level convolution models: tdrelconv, tdrelconvlp,
and trdelconvlpext (as with the relline case, they were added and removed
in XSPEC 12.10.1 but are not exposed to the users as model classes).