mpl jointplot now respects the ratio argument (top plot is larger)
With this change the sherpa[.astro].ui.plot_fit_resid/delchi routines
now display with the top plot being twice the height of the bottom
plot when using the matplotlib (pylab) backend.

There is no test for this as there is currently no good test suite
for the plot contents when a plotting backend is available.