Matplotlib and chips backends disagree on the number of subplots
It is not obvious whether jointplot is meant to be used with anything
but a single column: the chips backend seems to suggest it does, but
then only creates a single column anyway, and the matplotlib backend
just ignores the number-of-columns argument (at least in set_jointplot).

Follow the matplotlib code here.