Supply a default option for all configuration-value requests
The fallback option was not supported for the config.get method in
Python 2.7, but we can take advantage of it here. This does not
really change anything - although it does clean up the code slightly -
but providing a default value makes it easier for a user switching
between Sherpa versions where configuration options may have been
added or deleted.

It does imply a coupling between the code and the configuration
file (in that we would expect the fallback value to match the
default value in our configuration files), but I think this cost
is worth it.