Remove the check on the color value (hex conversion)
As noted by @hamogu, this function does not appear to be well defined,
since it converts '0xfff' to '#000fff', which is not correct. This
routine seems to be copied from one in the ChIPS backend (very-slightly
modified), but it is not clear why either are needed (since ChIPS
doesn't accept string hex values). As there isn't an obvious use case,
I have decided to remove it.