Expand the primini test (for when #431 has been fixed)
This adds validation of some pre- and post-conditions for the
primini fit. Note that the post-fit values come from the proposed fix,
so are not independently evaluated: they should be considered as
a regression test, and check that something happened during the
primini fit. Any validation of the primini algorithm is assumed to
have been done elsewhere.