Avoid bug #431 (failure when fitting with primini iteration method)
The code fix (to handle any None values returned by get_staterror)
looks sensible, but I can not see why the old code ever worked.
Looking at the commit that caused the change in behavior, there is
no obvious change to the get_staterror code path, and I have not
looked at this code recently enough to make sense of it.

This is a long way of saying that while this looks to me a reasonable
fix, there could be odd things still lurking around.