test_astro_ui_plot: add tests for plotting
Add extra tests to the astro ui plotting tests to check that we can
call many of the plot routines when there is a PHA data set or
an image one (in which case the contour routines are exercised).
These are very-limited tests, in that they relly just check that the
routine can be called.

There are also some matplotlib-specific tests of plot_data and
plot_model, which actually check the output plot (by inspecting the
matplotlib object hierarchy and checking for a limited set of values).
These tests are directed at checking that the defaults can be changed,
and are aimed at supporting any needed changes for matplotlib 3.3
(as matplotlib 3.1 currently creates warning messages when model-related
plots are created).

These are matplotlib-specific tests and check that we can change the
plot defaults and see the results. This is one of the few existing
tests that actually check the plots created by the tests (by querying
the state of the matplotlib objects).