Allow plot preferences to be over-ridden by the user when calling plot
The aim is to support the ui layer being able to say

    ui.plot_fit(color='green')

and having this override the plot preferences *for this call only*.

The implementation is surprisingly simple, although there are issues
on how joint plots are to be handled, ui-layer changes to pass this
through, and backend changes.

The method used to merge preferences and user settings is set up to
allow sub-classes to over-ride this (e.g. to improve handling of joint
plots), but it turns out that this logic is actually handled in the
ui layer, so the multiple "_merge_settings" calls could be removed.

There is no documentation yet.