Confidence2D.contour method takes overplot and not overcontour
It looks like the contour method of sherpa.plot.Confidence2D should
really accept an overcontour argument rather than overplot. This
commit supports both (so that the overplot argument can be removed
at a later time if this is a good idea).

The idea is that the "overplot/contour" action will be set if either
overplot or overcontour is needed. Without this change the code would
error out if a user tried to call an overcontour() method with a
confusing error message:

    TypeError: contour() got multiple values for keyword argument 'overcontour'

There is no documentation change, but should be.