bugfix: handle point preferences in region plots
Rather than send the Confidence2D object to the point class, just
create a temporary point object, since otherwise the merge of the
preferences with data was being done with the wrong set of preferences.
An alternative approach would have been to look at handling this
in the _merge_settings method, but this is simpler.