fix: plot_source/model_component
The plot component classes use multiple inheritance, and it turns out
that the default merge they use picks up the wrong preference settings.
In a different commit in this PR a similar problem was avoided by
creating a temporary object (for a point), but in this case I over-ride
the merge-settings method to pick out the correct preference settings.