Docs: improve low-level plot documentation
The aim was to add in information about passing keyword arguments to
the backend, but this required adding a bunch of documentation first.

The aim was to ensure over-ridden methods (e.g. plot and prepare)
were documented correctly, since the supported arguments change as
you go through sub-classes (e.g. DataPlot compared to Plot).

The contour methods and classes have not been updated.