Refactor passing of info to the plot prepare methods
The aim is to make it safer to pass arguments to the prepare method
of a plot when using the UI layer, and to make it easier to add
additional arguments in the future. There is no increase in functionality
in this commit.

The main change is that we now use

    self._plot(id, self._orderplot, prepare_args={'orders': orders},
               replot=replot, ..., **kwargs)

rather than

    self._plot(id, self._orderplot, None, None, None, None, orders,
               replot=replot, ..., **kwargs)

It would also be possible to package up the "replot=..., **kwargs"
arguments into a named argument plot_args, to match prepare_args,
but I have decided not to go that far here, even though it would make
things a bit clearer.