Initial support for rescaling the background data in a plot
The background data plot can now display its data re-scaled to match
a "source" aperture. Since the DataPHA object only maps from source to
background, so there is no link from a background dataset to its "source",
the source has to be manually specified.

There is currently a requirement that the source and background
have the same filtering and grouping, but this would ideally be
relaxed, since you may want to use different settings when fitting
a model to the background.

Tests have been added, but a few cases have not been covered, namely
the y error bars, grouped data, and several possible "error" conditions.

This change has only been applied to the backgroud "data" plot, but
ideally it would be applied to other background-plots too: for instance
if you have fit a model to the background, you may want to see the
model prediction scaled to match the source observation.

The implementation is not ideal, since it copoes a lot of object-specific
knowledge - in particular the workings of the DataPHA class - into the
plot class structure. I think it would be better if the DataPHA class,
could support this directly. There is the sum_background_data method but
this combines all the background data sets and it is unclear about
what its semantics are (e.g. with filtering and grouping).