Add the rescale argument to plot_bkg/get_bkg_plot
The background can now be rescaled to match the source aperture from
the sherpa.astro.ui layer by setting the rescale argument in plot_bkg
or get_bkg_plot to True. The aim is to allow a sequence like

    plot_data()
    plot_bkg(rescale=True)

so that user's can easily see how the background compares to the
source data.

This is only supported by plot_bkg: it should be added to other plots,
such as plot_bkg_model and plot_bkg_fit.