resid/ratio/delchi plots always use a linear scale for the y axis
Address #586

The plot method for these classes has been over-ridden to force the
ylog preference setting to be False, and to drop the ylog keyword
argument if specified. This means that a command like (from the
ui layer)

    set_ylog()
    plot_fit_resid()

will use a logarithmic Y axis for the top plot but a linear one for
the bottom plot.

This might be better done as a Mixin / Multi-Inheritance class, since
the logic is the same for all three cases, but I have not looked into
this approach.