Add a to-be-filled test suite for multiple ARF/RMF responses
The aim is to provide some testing of PHA fitting with multiple
responses, such as those provided by load_multi_arfs/load_multi_rmfs.

The initial test is for multiple ARFs only, and uses set_arf
rather than load_multi_arfs, so that this can be done without any
files needing to be added to the repository.