Add 1D regrid tests based on issue #722
This adds a basic test of what happens when there is partial overlap
of the "data" grid and the requested grid for 1D data (both
non-integrated and integrated). Tests include checking what
happens when there is a mis-match in bin sizes (both smaller and
larger). The tests have been adjusted as the fixes for #722 have
been made (see #747).

This is a simplified test since the model evaluates to 0 where the
data and requested grids do not overlap (by construction).

Tolerances have been adjusted so that the tests pass on my
machines.