Add basic tests where regrid covers the data space (1D)
Previous tests were added to show what happens when the regrid grid
is smaller than the output grid, these test the opposite.

Note that the 1D non-integrated case fails for a large bin width, as
the last few bins are >> 0 for some reason. The test has been marked
as xfail because of this.

The 1D integrated case appears to be working correctly (although there's
an interesting pattern if you look at abs(expected - got) where the
differences seem to alternate between 0 and ~1e-16. This is not
explored in these tests but was noted when manually checking the
results.