Add a test or rebin with partial overlap
This test is added as a check for #722, and is intended to show that
rebin works okay when the bins do not line up perfectly, but
the requested grid is larger than the input grid (in this case with 0's
at the edge of the grid to simplify things). This is to show that the
problems in #722 are with the regrid code, not rebin.