BUGFIX: Python 3.8 TypeError about integer scalar arrays
The error message was:

TypeError: only integer scalar arrays can be converted to a scalar index

It looks like Python 3.8.1 / NumPy 1.18.1 doesn't want to automatically convert
a NumPy int64 array (with one element) to a Python integer type.
It's not clear to me whether this is an issue with one, both, or a combo
of these two versions.

The approach taken is to explicitly convert the integer value to a
Python one: this conversion works if brightPixel is a scalar or a one-element
NumPy array, which holds here.