Minor clean up of sherpa.utils.is_binary_file
A note has been added to the docstring to be more explicit about what
"binary" means, and the internal documentation has been made more
explicit (rather than just comment on Python 2 vs 3 code paths).

The code has been changed to make it more obvious where the
errors are expected and using a context manager to handle the
file closing. There should be no functional changes to users of
this function.