Remove python 2 vs 3 comment
Be explicit in the comments about the attempted string conversion;
this removes the note that this was added in the python 2 to 3
update.

We could update the documentation to state the parameter names are
required to be strings and avoid this, but for now this is the
less invasive change.