Change datastack keyword query to support missing keys
Change query_by_header_keyword so that it skips data sets which
do not contain the keyword, rather than raising a KeyError. I
feel that it makes more sense to skip files rather than error out,
and I feel that the original code supports this view: it skipped
dataset with no header, and did have a check for missing keys, but
this was done after retrieving the key and so ended up never being
used.

This also removes support for header keywords stored as byte strings,
since neither AstroPy or Crates support this (and, although the FITS
standard is not written using Python terminology, I would argue
that it is intended that key values are strings and not byte strings).
There was a period during development when crates did generate
byte strings, but this never made it to a release).